[#43] simplify and rename api-transport
