+++ components/RayDumpPreview/RayTypesPreview/LockPayload.vue
@@ -1,18 +1,32 @@
 <template>
   <div class="ray-type-lock">
-    <button :disabled="disabled" @click="continueExecution"
-            class="btn btn--continue active:bg-grey-300">
-            <span class="w-3 h-3 block">
-                <svg xmlns="http://www.w3.org/2000/svg" width="100%" height="100%" viewBox="0 0 20 20">
-                    <path fill="green" fill-rule="evenodd"
-                          d="M16.75 10.83L4.55 19A1 1 0 0 1 3 18.13V1.87A1 1 0 0 1 4.55 1l12.2 8.13a1 1 0 0 1 0 1.7z"/>
-                </svg>
-            </span>
+    <button
+      :disabled="disabled"
+      class="btn btn--continue active:bg-grey-300"
+      @click="continueExecution"
+    >
+      <span class="w-3 h-3 block">
+        <svg
+          xmlns="http://www.w3.org/2000/svg"
+          width="100%"
+          height="100%"
+          viewBox="0 0 20 20"
+        >
+          <path
+            fill="green"
+            fill-rule="evenodd"
+            d="M16.75 10.83L4.55 19A1 1 0 0 1 3 18.13V1.87A1 1 0 0 1 4.55 1l12.2 8.13a1 1 0 0 1 0 1.7z"
+          />
+        </svg>
+      </span>
 
       <span>Continue</span>
     </button>
-    <button :disabled="disabled" @click="stopExecution"
-            class="btn btn--stop active:bg-grey-300">
+    <button
+      :disabled="disabled"
+      class="btn btn--stop active:bg-grey-300"
+      @click="stopExecution"
+    >
       <span class="w-3 h-3 bg-red-700 block"></span>
       <span>Stop execution</span>
     </button>
@@ -20,14 +34,11 @@
 </template>
 
 <script lang="ts">
-import {defineComponent, PropType} from "vue";
-import {RayPayload} from "~/config/types";
-import {apiTransport} from '~/utils/events-transport'
+import { defineComponent, PropType } from "vue";
+import { RayPayload } from "~/config/types";
+import { apiTransport } from "~/utils/api-transport";
 
-const {
-  rayStopExecution,
-  rayContinueExecution
-} = apiTransport({onEventReceiveCb: () => {}})
+const { rayStopExecution, rayContinueExecution } = apiTransport();
 
 export default defineComponent({
   props: {
@@ -39,23 +50,23 @@ export default defineComponent({
   data() {
     return {
       disabled: false,
-    }
-  },
-  methods: {
-    continueExecution() {
-      this.disabled = true
-      rayContinueExecution(this.hash)
-    },
-    stopExecution() {
-      this.disabled = true
-      rayStopExecution(this.hash)
-    }
+    };
   },
   computed: {
     hash() {
-      return this.payload.content.name
-    }
-  }
+      return this.payload.content.name;
+    },
+  },
+  methods: {
+    continueExecution() {
+      this.disabled = true;
+      rayContinueExecution(this.hash);
+    },
+    stopExecution() {
+      this.disabled = true;
+      rayStopExecution(this.hash);
+    },
+  },
 });
 </script>
