+++ utils/io/types.ts
@@ -3,6 +3,5 @@ import { ServerEvent } from "~/config/types";
 export type LoggerParams = [string, unknown]
 
 export interface ApiConnection {
-  onEventReceiveCb: (param: ServerEvent<unknown>) => void
   loggerCb?: (params: LoggerParams) => void
 }
